requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
python-dotenv
openai
tiktoken
google-generativeai